// app/admin/collections/interfaces.ts
export interface Product {
    id: string;
    name: string;
    slug: string;
    price: string;
    inventory: number;
    isActive: boolean;
    categoryId: string | null;
    categoryName: string | null;
}

export interface Collection {
    id: string;
    name: string;
}

export const PRODUCTS_PER_PAGE = 10;